// src/interfaces/Product.ts
export interface IProduct {
  _id: string
  vendor_id: string
  sku: string
  name: string
  is_active: boolean
  description: string
  notes: string
  times_sold: number
  created_by: string
  updated_by: string
  back_order: string[]
  on_sale: boolean
  family_id: string
  family_name: string
  stock: number
  brand: string
  category: string
  category_sub1: string
  category_sub2: string
  category_id: string
  category_slug: string
  msrp: number
  each_weight: string
  each_price: number
  each_length: string
  each_width: string
  each_height: string
  each_cost: number
  each_discount: number
  each_upc: string
  case_qty: number
  case_price: number
  case_cost: number
  case_weight: string
  case_length: string
  case_width: string
  case_height: string
  case_upc: string
  case_discount: number
  pallet_qty: number
  pallet_price: number
  pallet_cost: number
  pallet_weight: string
  pallet_length: string
  pallet_width: string
  pallet_height: string
  pallet_upc: string
  pallet_discount: number
  warranty: string
  hazmat: boolean
  legacy_sku: string
  image_thumb: string
  image_medium: string
  image_family_thumb: string
  image_family_medium: string
  sold_in_quantities_of: number
  discontinued: boolean
}